+++ src/main/java/domain/ladder/Paths.java
@@ -1,6 +1,5 @@
 package domain.ladder;
 
-import constant.PathStatus;
 import domain.LineNumber;
 
 import java.util.ArrayList;
